Refactor DI for preferences as it is AppScope DI
Also move bloc initial tasks loading event to its constructor,
it's more suitable place for it.
