+++ lib/src/feature/app/app_entry_point.dart
@@ -9,8 +9,8 @@ import '../../core/api/interceptors/auth_interceptor.dart';
 import '../../core/database/database_impl.dart';
 import '../../core/router/router.dart';
 import '../../core/uikit/uikit.dart';
+import 'app_settings.dart';
 import 'di/app_scope.dart';
-import 'preferences.dart';
 
 /// The entry point of the app
 class YetAnotherTodoApp extends StatelessWidget {
@@ -24,10 +24,11 @@ class YetAnotherTodoApp extends StatelessWidget {
     super.key,
   }) : _appRouter = AppRouter();
 
-  // All the DI will be done in the [YetAnotherTodoApp] class
   @override
   Widget build(BuildContext context) {
     return AppScope(
+      // ------ Main App Scope Init ------
+      db: db,
       dio: AppDioConfigurator.create(
         interceptors: [
           AuthDioInterceptor(
@@ -36,24 +37,22 @@ class YetAnotherTodoApp extends StatelessWidget {
         ],
         url: dotenv.env['API_BASE_URL']!,
       ),
-      db: db,
-      child: AppPreferencesScope(
-        preferences: sharedPrefs,
-        child: MaterialApp.router(
-          routerConfig: _appRouter.config(),
-          debugShowCheckedModeBanner: false,
-          theme: AppThemeData.lightTheme,
-          darkTheme: AppThemeData.darkTheme,
-          // Override the default localization delegate
-          // so the app is consistent with its language.
-          localizationsDelegates: const [
-            GlobalMaterialLocalizations.delegate,
-            GlobalWidgetsLocalizations.delegate,
-            GlobalCupertinoLocalizations.delegate,
-            AppLocalizations.delegate,
-          ],
-          supportedLocales: AppLocalizations.supportedLocales,
-        ),
+      appSettingsRepository: AppSettingsRepository(
+        sharedPreferences: sharedPrefs,
+      ),
+      // ---------------------------------
+      child: MaterialApp.router(
+        routerConfig: _appRouter.config(),
+        debugShowCheckedModeBanner: false,
+        theme: AppThemeData.lightTheme,
+        darkTheme: AppThemeData.darkTheme,
+        localizationsDelegates: const [
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+          GlobalCupertinoLocalizations.delegate,
+          AppLocalizations.delegate,
+        ],
+        supportedLocales: AppLocalizations.supportedLocales,
       ),
     );
   }
