+++ lib/src/feature/app/di/app_scope.dart
@@ -1,16 +1,17 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import '../../../core/database/database_impl.dart';
-import '../../../core/tools/logger.dart';
 import '../../todo/bloc/todo_bloc/todo_bloc.dart';
 import '../../todo/data/repository/api_todo_repository_impl.dart';
 import '../../todo/data/repository/db_todo_repository_impl.dart';
 import '../../todo/domain/repository/api_todo_repository.dart';
 import '../../todo/domain/repository/db_todo_repository.dart';
+import '../app_settings.dart';
 
 class AppScope extends InheritedWidget {
   final TodoRepositoryApi _todoApiRepository;
   final TodoRepositoryDb _todoDbRepository;
+  final AppSettingsRepository _appSettingsRepository;
 
   late final TodoBloc todoBloc;
 
@@ -18,15 +19,15 @@ class AppScope extends InheritedWidget {
     required super.child,
     required Dio dio,
     required AppDatabaseImpl db,
+    required AppSettingsRepository appSettingsRepository,
     super.key,
   })  : _todoApiRepository = TodoRepositoryApiImpl(baseDioClient: dio),
-        _todoDbRepository = TodoRepositoryDbImpl(db) {
+        _todoDbRepository = TodoRepositoryDbImpl(db),
+        _appSettingsRepository = appSettingsRepository {
     todoBloc = TodoBloc(
       todoRepositoryApi: _todoApiRepository,
       todoRepositoryDb: _todoDbRepository,
-    )..add(const TodoEvent.loadTodos());
-
-    logger.i('App scope initialized');
+    );
   }
 
   @override
@@ -41,4 +42,10 @@ class AppScope extends InheritedWidget {
   static AppScope? maybeOf(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<AppScope>();
   }
+}
+
+extension AppScopeX on BuildContext {
+  AppScope get appScope => AppScope.of(this);
+
+  AppSettingsRepository get settings => appScope._appSettingsRepository;
 }
